requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
huggingface-hub==0.25.0
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
multidict==6.1.0
networkx==3.4.2
numpy==1.23.5
packaging==24.1
pandas==2.2.2
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
propcache==0.3.0
protobuf==5.28.2
psutil==6.0.0
pyparsing==3.1.4
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests @ file:///croot/requests_1721410876868/work
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==1.0.9
torch==1.13.0
torch-geometric==2.6.1
torchaudio==0.13.0
torchmetrics==1.4.2
torchvision==0.14.0
tqdm==4.66.5
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
wandb==0.18.1
yarl==1.18.3
